third_party/WebKit/LayoutTests/editing/execCommand/insert_list_at_end_of_paragraph.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>',
                '^This paragraph should be in a list.<br>|',
                'This paragraph should not be in a list.<br>',
            '</div>',
        ].join(''),
        'InsertUnorderedList',
        [
            '<div contenteditable>',
                '<ul><li>^This paragraph should be in a list.|</li></ul>',
                'This paragraph should not be in a list.<br>',
            '</div>',
        ].join('')),
        'InsertUnorderedList on a selection that ends at the start of a paragraph.');
</script>
